Separated powder related functions
